fix nits and add fromByte function to HashType

+++ src/main/scala/org/bitcoins/core/crypto/TransactionSignatureChecker.scala
@@ -43,7 +43,7 @@ trait TransactionSignatureChecker extends BitcoinSLogger {
       logger.error("Signature did not have a low s value")
       ScriptValidationFailureHighSValue
     } else if (ScriptFlagUtil.requireStrictEncoding(flags) && signature.bytes.nonEmpty &&
-      !HashType.hashTypes.contains(HashType(Int32(signature.bytes.last)))) {
+      !HashType.hashTypes.contains(HashType(signature.bytes.last))) {
       logger.error("signature: " + signature.bytes)
       logger.error("Hash type was not defined on the signature")
       ScriptValidationFailureHashType
